Tweaking changes to match the tutorial
